+++ embetter/vision/__init__.py
@@ -1,6 +1,11 @@
 from embetter.error import NotInstalled
+from embetter.vision._grab import ImageGrabber
+
 
 try:
     from embetter.vision._torchvis import TorchVision
 except ModuleNotFoundError:
     TorchVision = NotInstalled("TorchVision", "torch")
+
+
+__all__ = ["ImageGrabber"]
